Remove network thread (to be reworked). Fix switching server.

+++ src/Controls/TabBar/ControlsTab.cpp
@@ -75,8 +75,8 @@ void ControlsTab::reSizeTo(RECT & rc)
 bool ControlsTab::renameTab(const wchar_t *internalName, const wchar_t *newName)
 {
   bool foundIt = false;
-  size_t i = 0;
-  for ( ; i < _pWinVector->size() ; i++)
+  int i = 0;
+  for ( ; i < static_cast<int> (_pWinVector->size()) ; i++)
   {
     if ((*_pWinVector)[i]._internalName == internalName)
     {
